Rename projection_pitch -> pitch
