+++ include/fullscore/models/pitch.h
