+++ src/models/pitch.cpp
@@ -1,7 +1,7 @@
 
 
 
-#include <fullscore/models/projection_pitch.h>
+#include <fullscore/models/pitch.h>
